Attempt to initialize the app to provide an example to newcomers
The calculation isn't running when it's supposed to; the call to reevaluate in the calculator happens, but the deck appears to be empty when it does.
The call to reevaluate on deck import happens, except on the initialization import, and the cards value doesn't seem to be updating on time.